refactor: helper functions Types -> Utils module
